iceberg: move test utility into avro_utils
A previous commit added a test that defined some Avro serialization
utilites. These will be useful outside of tests, since it's generally
common for other parts of Redpanda to use iobufs as the in-memory buffer
of choice.
